requirements.txt
streamlit
google-generativeai
python-dotenv
langchain
langchain_experimental
langchain-openai
openai
Streamlit==1.34.0
beautifulsoup4
lxml
pypdf
PyPDF2
Flask